Moved DATA_INTERVAL and useRecurringCall from backups to shared/core
