+++ public/app/percona/backup/components/BackupInventory/BackupInventory.constants.ts
@@ -1,4 +1,3 @@
-export const DATA_INTERVAL = 5000;
 export const LIST_ARTIFACTS_CANCEL_TOKEN = 'listArtifacts';
 export const BACKUP_CANCEL_TOKEN = 'backup';
 export const RESTORE_CANCEL_TOKEN = 'restore';
